Decoupling user authentication API and create user API from Site Admin service. Created new User service and controller for user-centered API's during regisration and login.
